# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gigastt REST API
  version: "1.0"
  description: |
    Offline and streaming speech-to-text REST API powered by GigaAM v3 (rnnt head by default).
    On-device Russian speech recognition via ONNX Runtime.

    ## Endpoints

    - **Health & discovery**: `GET /health`, `GET /ready`, `GET /v1/models`
    - **File transcription**: `POST /v1/transcribe` — upload audio, receive full transcript
    - **Streaming transcription**: `POST /v1/transcribe/stream` — upload audio, receive SSE stream of partial/final segments
    - **OpenAI-compatible**: `POST /v1/audio/transcriptions` — multipart `file` + `model` → `{"text":"..."}`
    - **Async jobs** (opt-in `--enable-jobs`): `POST /v1/jobs`, `GET/DELETE /v1/jobs/{id}`, result + SSE events
    - **Admin**: `POST /v1/admin/reload` — loopback-only model hot-reload
    - **WebSocket** (see `docs/asyncapi.yaml`): `GET /v1/ws`

    Prometheus metrics are optional (behind `--metrics`) and exposed on a
    separate listener (default `127.0.0.1:9090`, override via `--metrics-listen`),
    not on this API surface.

    ## Audio Formats

    Supported upload formats: WAV, MP3, M4A/AAC, OGG/Vorbis, OGG/Opus (`.opus`),
    WebM/Opus (what a browser's `MediaRecorder` produces), FLAC. Raw headerless
    telephony streams are accepted with
    `?codec=pcmu|pcma|g722&sample_rate=N` (see the transcribe parameters).
    Containers are decoded server-side via [symphonia](https://github.com/pdeljanov/Symphonia)
...

    ## Limits

...
servers:
...
paths:
...
  /v1/audio/transcriptions:
    post:
      summary: OpenAI-compatible file transcription
      description: |
        OpenAI Audio Transcriptions compatible endpoint for clients that speak
        that API shape (llama-swap, Hermes Agent, OpenAI SDKs with a custom
        `base_url`). Multipart form with required `file` and optional
        `model` / `response_format` / `language` / `timestamp_granularities[]`.
        `model` is accepted for compatibility and ignored (loaded engine always
        used). Same audio formats and duration limits as `/v1/transcribe` (no
        default cap; see `--max-audio-secs`).

        `response_format`: `json` (default, `{"text"}`) · `text` · `srt` · `vtt`
        · `verbose_json` (Whisper-style: `task`, `language`, `duration`, `text`,
        optional `segments`/`words`). For diarization, telephony codecs, or
        native knobs use `POST /v1/transcribe`.
      operationId: createTranscription
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to transcribe.
                model:
                  type: string
                  description: |
                    Accepted for OpenAI client compatibility; ignored. Any string
                    works (`whisper-1`, a local alias, or a gigastt model id).
                  example: whisper-1
                response_format:
                  type: string
                  enum: [json, text, srt, vtt, verbose_json]
                  default: json
                  description: Output shape (OpenAI Whisper formats).
                language:
                  type: string
                  description: |
                    Echoed in `verbose_json` (default `ru`). Does not switch the
                    loaded recognition head.
                  example: ru
                timestamp_granularities:
                  type: array
                  items:
                    type: string
                    enum: [word, segment]
                  description: |
                    With `verbose_json`: include `words` and/or `segments`.
                    Form field name is `timestamp_granularities[]`. Default when
                    neither is set: segments only.
                prompt:
                  type: string
                  description: Accepted and ignored (SDK compatibility).
                temperature:
                  type: string
                  description: Accepted and ignored (SDK compatibility).
                stream:
                  type: boolean
                  default: false
                  description: |
                    When true, respond with SSE (`text/event-stream`) of
                    `transcript.text.delta` events, a final
                    `transcript.text.done`, and `data: [DONE]`. Only valid with
                    `response_format` json or text (default).
      responses:
        "200":
          description: |
            Transcription result. Content type depends on `response_format` /
            `stream`: JSON for `json`/`verbose_json`, `text/plain` for `text`,
            captions for `srt`/`vtt`, or `text/event-stream` when `stream=true`.
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                data: {"type":"transcript.text.delta","delta":"Привет"}

                data: {"type":"transcript.text.done","text":"Привет"}

                data: [DONE]
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [text]
                    properties:
                      text:
                        type: string
                        example: "Привет, как дела?"
                  - type: object
                    required: [task, language, duration, text]
                    properties:
                      task:
                        type: string
                        example: transcribe
                      language:
                        type: string
                        example: ru
                      duration:
                        type: number
                        format: float
                      text:
                        type: string
                      segments:
                        type: array
                        items:
                          type: object
                      words:
                        type: array
                        items:
                          type: object
              examples:
                json:
                  value:
                    text: "Привет, как дела?"
                verbose_json:
                  value:
                    task: transcribe
                    language: ru
                    duration: 3.5
                    text: "Привет, как дела?"
                    segments:
                      - id: 0
                        seek: 0
                        start: 0.5
                        end: 3.5
                        text: " Привет, как дела?"
            text/plain:
              schema:
                type: string
                example: "Привет, как дела?"
            application/x-subrip:
              schema:
                type: string
            text/vtt:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/TranscriptionError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          example: "Request body exceeds the configured size limit"
          description: Human-readable error description
        code:
          type: string
          example: payload_too_large
          description: Machine-readable error code
        retry_after_ms:
          type: integer
          example: 30000
          description: |
            Suggested retry delay in milliseconds. Present only for transient
            backpressure errors (pool saturation).
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Empty request body"
            code: empty_body
    PayloadTooLarge:
      description: |
        Request body exceeds `--body-limit-bytes`, or decoded audio exceeds a
        configured duration limit (`--max-audio-secs`, or the fixed
        ~30-minute safety ceiling on whole-buffer paths: diarization,
        `channels=split`, telephony).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            payload_too_large:
              summary: Body exceeds --body-limit-bytes
              value:
                error: "Request body exceeds the configured size limit"
                code: payload_too_large
            audio_too_long:
              summary: Decoded audio exceeds a configured duration limit
              value:
                error: "Audio file too long. Maximum supported: 1800s."
                code: audio_too_long
    TranscriptionError:
      description: Transcription failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Transcription failed. Check audio format."
            code: transcription_error
...
    ServiceUnavailable:
      description: Server busy or shutting down
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            timeout:
              summary: Pool saturated
              value:
                error: "Server busy, try again later"
                code: timeout
                retry_after_ms: 30000
            pool_closed:
              summary: Graceful shutdown in progress
              value:
                error: "Server is shutting down"
                code: pool_closed
    TooManyRequests:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Too many requests"
            code: rate_limited
            retry_after_ms: 30000
    GatewayTimeout:
      description: Inference timed out
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Inference timed out."
            code: inference_timeout
...